fix: initialize redux in prod mode by default
We were accidentally initializing in dev mode in production.

Also cleaning up some other misc detritus.
