+++ openedx.yaml
@@ -1,6 +1,6 @@
 # This file describes this Open edX repo, as described in OEP-2:
 # http://open-edx-proposals.readthedocs.io/en/latest/oeps/oep-0002.html#specification
 
-nick: acct
+nick: prof
 oeps: {}
 owner: edx/arch-team
