+++ src/data/reducers.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import { userAccount } from '@edx/frontend-auth';
 
-import { reducer as profilePage } from './profile';
+import { reducer as profilePage } from '../profile';
 
 const createRootReducer = () =>
   combineReducers({
