+++ src/data/sagas.js
@@ -1,6 +1,6 @@
 import { all } from 'redux-saga/effects';
 
-import { saga as profileSaga } from './profile';
+import { saga as profileSaga } from '../profile';
 
 export default function* rootSaga() {
   yield all([
