+++ src/index.jsx
@@ -10,7 +10,7 @@ import Footer, { messages as footerMessages } from '@edx/frontend-component-foot
 
 import appMessages from './i18n';
 import { ProfilePage, NotFoundPage } from './profile';
-import configureStore from './store';
+import configureStore from './data/configureStore';
 
 import './index.scss';
 import './assets/favicon.ico';
